Refact: refact 검색창 컴포넌트 생성
